backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
beautifulsoup4>=4.12.2
requests>=2.31.0
langchain>=0.1.0
langchain-groq>=0.1.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
lxml>=4.9.3
